guard kernel moves

+++ build_kernel.sh
@@ -12,10 +12,14 @@ case $DEVICE in
 
         if [[ $DEVICE == "coral" ]]; then
             KBUILD_BUILD_VERSION=1 KBUILD_BUILD_USER=build-user KBUILD_BUILD_HOST=build-host KBUILD_BUILD_TIMESTAMP="Thu 01 Jan 1970 12:00:00 AM UTC" BUILD_CONFIG=private/msm-google/build.config.floral build/build.sh
-            rsync -av --delete out/android-msm-pixel-4.14/dist/ /opt/build/grapheneos/device/google/coral-kernel/
+            if [[ $SKIP_GRAPHENEOS == "false" ]]; then
+                rsync -av --delete out/android-msm-pixel-4.14/dist/ /opt/build/grapheneos/device/google/coral-kernel/
+            fi
         else
             KBUILD_BUILD_VERSION=1 KBUILD_BUILD_USER=build-user KBUILD_BUILD_HOST=build-host KBUILD_BUILD_TIMESTAMP="Thu 01 Jan 1970 12:00:00 AM UTC" BUILD_CONFIG=private/msm-google/build.config.sunfish build/build.sh
-            rsync -av --delete out/android-msm-pixel-4.14/dist/ /opt/build/grapheneos/device/google/sunfish-kernel/
+            if [[ $SKIP_GRAPHENEOS == "false" ]]; then
+                rsync -av --delete out/android-msm-pixel-4.14/dist/ /opt/build/grapheneos/device/google/sunfish-kernel/
+            fi
         fi
         ;;
     bramble|redfin|barbet)
@@ -24,7 +28,9 @@ case $DEVICE in
         repo init -u https://github.com/GrapheneOS/kernel_manifest-redbull.git -b refs/tags/"${MANIFEST}"
         repo sync -j${NPROC_SYNC} --force-sync --no-clone-bundle --no-tags
         BUILD_CONFIG=private/msm-google/build.config.redbull.vintf build/build.sh
-        rsync -av --delete out/android-msm-pixel-4.19/dist/ /opt/build/grapheneos/device/google/redbull-kernel/vintf/
+        if [[ $SKIP_GRAPHENEOS == "false" ]]; then
+            rsync -av --delete out/android-msm-pixel-4.19/dist/ /opt/build/grapheneos/device/google/redbull-kernel/vintf/
+        fi
         ;;
     oriole|raven)
         mkdir -p /opt/build/kernel/"${DEVICE}"
@@ -32,7 +38,9 @@ case $DEVICE in
         repo init -u https://github.com/GrapheneOS/kernel_manifest-raviole.git -b refs/tags/"${MANIFEST}"
         repo sync -j${NPROC_SYNC} --force-sync --no-clone-bundle --no-tags
         LTO=full BUILD_AOSP_KERNEL=1 ./build_slider.sh
-        rsync -av --delete out/mixed/dist/ /opt/build/grapheneos/device/google/raviole-kernel/
+        if [[ $SKIP_GRAPHENEOS == "false" ]]; then
+            rsync -av --delete out/mixed/dist/ /opt/build/grapheneos/device/google/raviole-kernel/
+        fi
         ;;
     bluejay)
         mkdir -p /opt/build/kernel/"${DEVICE}"
@@ -40,7 +48,9 @@ case $DEVICE in
         repo init -u https://github.com/GrapheneOS/kernel_manifest-bluejay.git -b refs/tags/"${MANIFEST}"
         repo sync -j${NPROC_SYNC} --force-sync --no-clone-bundle --no-tags
         LTO=full BUILD_AOSP_KERNEL=1 ./build_bluejay.sh
-        rsync -av --delete out/mixed/dist/ /opt/build/grapheneos/device/google/bluejay-kernel/
+        if [[ $SKIP_GRAPHENEOS == "false" ]]; then
+            rsync -av --delete out/mixed/dist/ /opt/build/grapheneos/device/google/bluejay-kernel/
+        fi
         ;;
     panther|cheetah)
         mkdir -p /opt/build/kernel/"${DEVICE}"
@@ -48,7 +58,9 @@ case $DEVICE in
         repo init -u https://github.com/GrapheneOS/kernel_manifest-pantah.git -b refs/tags/"${MANIFEST}"
         repo sync -j${NPROC_SYNC} --force-sync --no-clone-bundle --no-tags
         LTO=full BUILD_AOSP_KERNEL=1 ./build_cloudripper.sh
-        rsync -av --delete out/mixed/dist/ /opt/build/grapheneos/device/google/pantah-kernel/
+        if [[ $SKIP_GRAPHENEOS == "false" ]]; then
+            rsync -av --delete out/mixed/dist/ /opt/build/grapheneos/device/google/pantah-kernel/
+        fi
         ;;
     lynx)
         mkdir -p /opt/build/kernel/"${DEVICE}"
@@ -56,6 +68,8 @@ case $DEVICE in
         repo init -u https://github.com/GrapheneOS/kernel_manifest-lynx.git -b refs/tags/"${MANIFEST}"
         repo sync -j${NPROC_SYNC} --force-sync --no-clone-bundle --no-tags
         LTO=full BUILD_AOSP_KERNEL=1 ./build_lynx.sh
-        rsync -av --delete out/mixed/dist/ /opt/build/grapheneos/device/google/lynx-kernel/
+        if [[ $SKIP_GRAPHENEOS == "false" ]]; then
+            rsync -av --delete out/mixed/dist/ /opt/build/grapheneos/device/google/lynx-kernel/
+        fi
         ;;
 esac
